# api.yaml
swagger: "2.0.0"
info:
  title: ShopOnAir
  description: Api documentation for the shoponair website
  version: 1.0.0
  contact:
    email: choudharysidharth082000@gmail.com
  servers:
    - url: "http://localhost:3000"
      description: "Local Server Running"
    - url: "https://shoponairbackend.herokuapp.com"
      description: "Main Server"
schemes: [http, https]
paths:
...
  /v1/profile/profileGet/{userID}:
    get:
      tags:
        - name: User Profile
      name: Getting user Profile By ID
      description: lorem
      parameters:
        - name: userID
          in: path
          type: string
          required: true
      responses:
        200:
          description: OK
...
  /v1/profile/deleteProfile/{userID}:
    delete:
      tags:
        - name: User Profile
      name: Delete the user Profile
      description: This will delete the user Profile of the user
      parameters:
        - name: userID
          in: path
          type: string
          required: true
      responses:
        200:
          description: OK
...
  /v1/profile/addprofile/{userID}:
    post:
      tags:
        - name: User Profile
      name: Add User Profile
      description: Adding the user Profile if present
      consumes:
        multipart/formdata
      produces:
        application/json
      parameters:
        - name: userID
          in: path
          type: string
          required: true
        - name: name
          in: formData
          type: string
          required: true
        - name: mobileNumber
          in: formData
          type: string
          required: true
        - name: facebookID
          in: formData
          type: string
          required: true
        - name: instagramID
          in: formData
          type: string
          required: true
        - name: whatsappNumber
          in: formData
          type: string
          required: true
        - name: profileImage
          in: formData
          type: file
          required: true
      responses:
        200:
          description: OK
  /v1/profile/editProfile/{userID}:
    put:
      tags:
        - name: User Profile
      name: Edit User Profile
      description: Edtiing the user Profile if present
      consumes:
        multipart/formdata
      produces:
        application/json
      parameters:
        - name: userID
          in: path
          type: string
          required: true
        - name: name
          in: formData
          type: string
          required: true
        - name: mobileNumber
          in: formData
          type: string
          required: true
        - name: facebookID
          in: formData
          type: string
          required: true
        - name: instagramID
          in: formData
          type: string
          required: true
        - name: whatsappNumber
          in: formData
          type: string
          required: true
        - name: profileImage
          in: formData
          type: file
          required: true
      responses:
        200:
          description: OK
  # /v1/profile/editProfile/{userID}:
  #   put:
  #     tags:
  #       - name: User Profile
  #     name: Edits the users Profile
  #     description: This api will edit the users profile
  #     parameters:
  #       - name: userID
  #         in: path
  #         type: string
  #         required: true
  #       - name: name
  #         in: formData
  #         type: string
  #         required: true
  #       - name: mobileNumber
  #         in: formData
  #         type: string
  #         required: true
  #       - name: facebookID
  #         in: formData
  #         type: string
  #         required: true
  #       - name: instagramID
  #         in: formData
  #         type: string
  #         required: true
  #       - name: whatsappNumber
  #         in: formData
  #         type: string
  #         required: true
  #       - name: profileImage
  #         in: formData
  #         type: file
  #         required: true
  #     responses:
  #       200:
  #         description: OK
...